requirements.txt
schedule
pymongo
boto3
Pillow